Day 1
Digging through nms to figure out methods and and how to override.

Created basic functionality of increasing/decreasing player range as well as changing the mob type.

Seems that the injection process is not working correctly, its setting the fields due to the b() method call and the NBT data however the CustomMobSpawner does not inject and persist.