Reflection
Turns out that the b() and a() methods don't actually store the MobSpawnerAbstract into the Tile but just set the Tile NBT Data to the Spawners NBT data. This is why the fields were being set such as mobName etc but the events weren't triggering. After looking deeper into the Tile class I realized that there isn't actually a setSpawner() method so reflection had to be used. After implementing reflection all events triggered and worked perfect.

- Removed Flames + Smoke when mobs spawn
- Firework now spawns whenever a mob spawns